Move svc ResourceLocation to 'inner' REST
Also move the test.
